[Refactor][Feat]

+++ src/main/java/com/whataburger/whataburgerproject/controller/dto/OptionReadResponseDto.java
@@ -13,5 +13,5 @@ public class OptionReadResponseDto {
     private Long optionId;
     private String optionName;
     private String imageSource;
-    private int optionCalories;
+    private Double optionCalories;
 }
